// sdk/nodejs/ec2/enums.ts
export const InstancePlatform = {
    LinuxUnix: "Linux/UNIX",
    RedHatEnterpriseLinux: "Red Hat Enterprise Linux",
    SuseLinux: "SUSE Linux",
    Windows: "Windows",
    WindowsWithSqlServer: "Windows with SQL Server",
    WindowsWithSqlServerEnterprise: "Windows with SQL Server Enterprise",
    WindowsWithSqlServerStandard: "Windows with SQL Server Standard",
    WindowsWithSqlServerWeb: "Windows with SQL Server Web",
} as const;

export type InstancePlatform = (typeof InstancePlatform)[keyof typeof InstancePlatform];